registry/etcd: use etcd as service registry
this CL adds a way to get hugerange keys by pagination, like:

Get(ctx, key, opts...)
  - option WithRev, specify the same revision as the one first Get(...) returns
  - option WithLimit, specify the number of keys returned one time
  - option WithRange, specify the end key if you could

From etcd issue list, we know: the hugerange requests may prevent the
write requests in earlier etcdv3 versions, it may lead to a problem,
that is the `applied index - commit index > 5000`, while 5000 is a
upperlimit, if this limit is hit, it means the etcd cannot quickly
finish the RAFT algorithm to keep the follower quickly follow up the
master. The etcd cluster is in a bad state. So etcd may drop some
requests to recover the state... Then, the client may encounter
keepalive error, and have to re-initialize the etcd client later.